dom/base/test/test_pluginMutedBeforePlay.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Mute window before the plugin starts playing</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<pre id="test">
</pre>
<iframe></iframe>

<script type="application/javascript">

// Copied from /dom/plugins/test/mochitest/utils.js
function getTestPlugin(pluginName) {
  var ph = SpecialPowers.Cc["@mozilla.org/plugin/host;1"]
                                 .getService(SpecialPowers.Ci.nsIPluginHost);
  var tags = ph.getPluginTags();
  var name = pluginName || "Test Plug-in";
  for (var tag of tags) {
    if (tag.name == name) {
      return tag;
    }
  }

  ok(false, "Could not find plugin tag with plugin name '" + name + "'");
  return null;
}
// Copied from /dom/plugins/test/mochitest/utils.js
function setTestPluginEnabledState(newEnabledState, pluginName) {
  var oldEnabledState = SpecialPowers.setTestPluginEnabledState(newEnabledState, pluginName);
  if (!oldEnabledState) {
    return;
  }
  var plugin = getTestPlugin(pluginName);
  while (plugin.enabledState != newEnabledState) {
    // Run a nested event loop to wait for the preference change to
    // propagate to the child. Yuck!
    SpecialPowers.Services.tm.currentThread.processNextEvent(true);
  }
  SimpleTest.registerCleanupFunction(function() {
    SpecialPowers.setTestPluginEnabledState(oldEnabledState, pluginName);
  });
}
setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);

SimpleTest.waitForExplicitFinish();

function runTest() {
  var iframe = document.querySelector("iframe");
  iframe.src = "file_pluginAudioNonAutoStart.html";

  function muteBeforePlay() {
    ok(!iframe.contentWindow.pluginMuted(), "Plugin should not be muted");
    iframe.contentWindow.toggleMuteState(true);

    iframe.contentWindow.startAudio();
    ok(iframe.contentWindow.pluginMuted(), "Plugin should still be muted after playing");

    iframe.contentWindow.stopAudio();
    SimpleTest.finish();
  }

  iframe.onload = function() {
    ok(true, "Already load iframe.");
    muteBeforePlay();
  }
}

onload = runTest;

</script>
</body>
</html>
